refactor: refactor login screen, kiwi background
