+++ frontend/src/components/login/screen.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { LoginBackground } from './background';
+import { LoginContainer } from './container';
 import { ReactComponent as TextLogoSvg } from './icons/text_logo.svg';
 
 const ContentContainer = styled.div`
@@ -7,6 +7,8 @@ const ContentContainer = styled.div`
   left: 0px;
   top: 0px;
   padding: 3rem 3rem;
+
+  width: 281px;
 `;
 
 const TextLogo = styled(TextLogoSvg)`
@@ -14,20 +16,7 @@ const TextLogo = styled(TextLogoSvg)`
   color: black;
 `;
 
-const Background = styled(LoginBackground)`
-  position: absolute;
-  z-index: -999999;
-
-  width: 100%;
-  height: 100%;
-`;
-
-const FormContainer = styled.div`
-  width: 281px;
-`;
-
-const Container = styled.div`
-  position: relative;
+const Container = styled(LoginContainer)`
   width: 100%;
   height: 100%;
 `;
@@ -36,12 +25,9 @@ export const LoginScreen = ({
   children,
 }: React.PropsWithChildren) => {
   return <Container>
-    <Background />
     <ContentContainer>
       <TextLogo />
-      <FormContainer>
-        {children}
-      </FormContainer>
+      {children}
     </ContentContainer>
   </Container>;
 };
